Crate fedimint_lnv2_client

Modules

api 🔒

cli 🔒

db 🔒

receive_sm 🔒

send_sm 🔒

Structs

LightningClientContext

LightningClientInit

LightningClientModule

ReceiveOperationMeta

SendOperationMeta

Enums

FinalReceiveOperationState

The final state of an operation receiving a payment over lightning.

FinalSendOperationState

The final state of an operation sending a payment over lightning.

LightningClientStateMachines

LightningOperationMeta

ReceiveError

ReceiveOperationState

The state of an operation receiving a payment over lightning.

SelectGatewayError

SendOperationState

The state of an operation sending a payment over lightning.

SendPaymentError

Constants

CONTRACT_CONFIRMATION_BUFFER 🔒

A two hour buffer in case either the client or gateway go offline

EXPIRATION_DELTA_LIMIT 🔒

Number of blocks until outgoing lightning contracts times out and user client can refund it unilaterally

Functions

generate_ephemeral_tweak 🔒

Type Aliases

ReceiveResult

SendResult