Struct fedimint_lnv2_client::LightningClientModule

pub struct LightningClientModule {
    pub federation_id: FederationId,
    pub cfg: LightningClientConfig,
    pub notifier: ModuleNotifier<LightningClientStateMachines>,
    pub client_ctx: ClientContext<Self>,
    pub module_api: DynModuleApi,
    pub keypair: KeyPair,
    pub admin_auth: Option<ApiAuth>,
    pub gateway_conn: Arc<dyn GatewayConnection + Send + Sync>,
}

Client side lightning module

Note that lightning gateways use a different version of client side module.

Fields
federation_id: FederationIdcfg: LightningClientConfignotifier: ModuleNotifier<LightningClientStateMachines>client_ctx: ClientContext<Self>module_api: DynModuleApikeypair: KeyPairadmin_auth: Option<ApiAuth>gateway_conn: Arc<dyn GatewayConnection + Send + Sync>
Implementations

impl LightningClientModule

pub async fn update_gateway_map(&self) -> !

This method updates the mapping from lightning node public keys to gateway api endpoints maintained in the module database once a day. When paying an invoice this enables the client to select the gateway that has created the invoice, if possible, such that the payment does not go over lightning, reducing fees and latency.

Client integrators are expected to call this function in a spawned task.

pub(crate) async fn select_gateway( &self, invoice: Option<Bolt11Invoice>, ) -> Result<(SafeUrl, RoutingInfo), SelectGatewayError>

pub(crate) async fn routing_info( &self, gateway_api: &SafeUrl, ) -> Result<Option<RoutingInfo>, GatewayConnectionError>

pub async fn send( &self, invoice: Bolt11Invoice, gateway_api: Option<SafeUrl>, ) -> Result<OperationId, SendPaymentError>

Pay an invoice. For testing you can optionally specify a gateway to route with, otherwise a gateway will be selected automatically.

pub async fn send_custom( &self, invoice: Bolt11Invoice, payment_fee_limit: PaymentFee, expiration_delta_limit: u64, gateway_api: Option<SafeUrl>, custom_meta: Value, ) -> Result<OperationId, SendPaymentError>

Pay an invoice. For testing you can optionally specify a gateway to route with, otherwise a gateway will be selected automatically.

pub(crate) async fn get_next_operation_id( &self, invoice: &Bolt11Invoice, ) -> Result<OperationId, SendPaymentError>

pub async fn subscribe_send( &self, operation_id: OperationId, ) -> Result<UpdateStreamOrOutcome<SendState>>

Subscribe to all updates of the send operation.

pub async fn await_send( &self, operation_id: OperationId, ) -> Result<FinalSendState>

Await the final state of the send operation.

pub async fn receive( &self, amount: Amount, gateway_api: Option<SafeUrl>, ) -> ReceiveResult

Request an invoice. For testing you can optionally specify a gateway to generate the invoice, otherwise a gateway will be selected automatically.

pub async fn receive_custom( &self, amount: Amount, expiry_secs: u32, description: Bolt11InvoiceDescription, payment_fee_limit: PaymentFee, gateway_api: Option<SafeUrl>, custom_meta: Value, ) -> Result<(Bolt11Invoice, OperationId), ReceiveError>

Request an invoice. For testing you can optionally specify a gateway to generate the invoice, otherwise a gateway will be selected automatically.

pub(crate) async fn create_contract_and_fetch_invoice( &self, recipient_static_pk: PublicKey, amount: Amount, expiry_secs: u32, description: Bolt11InvoiceDescription, payment_fee_limit: PaymentFee, gateway_api: Option<SafeUrl>, ) -> Result<(IncomingContract, Bolt11Invoice), ReceiveError>

Create an incoming contract locked to a public key derived from the recipient’s static module public key and fetches the corresponding invoice.

pub(crate) async fn receive_incoming_contract( &self, contract: IncomingContract, invoice: Bolt11Invoice, custom_meta: Value, ) -> Option<OperationId>

pub(crate) fn recover_contract_keys( &self, contract: &IncomingContract, ) -> Option<(KeyPair, AggregateDecryptionKey)>

pub async fn subscribe_receive( &self, operation_id: OperationId, ) -> Result<UpdateStreamOrOutcome<ReceiveState>>

Subscribe to all updates of the receive operation.

pub async fn await_receive( &self, operation_id: OperationId, ) -> Result<FinalReceiveState>

Await the final state of the receive operation.