Crate fedimint_lnv2_client

Modules

api

cli 🔒

db 🔒

receive_sm 🔒

send_sm 🔒

Structs

CreateBolt11InvoicePayload

LightningClientContext

LightningClientInit

LightningClientModule

Client side lightning module

PaymentFee

ReceiveOperationMeta

RoutingInfo

SendOperationMeta

SendPaymentPayload

Enums

Bolt11InvoiceDescription

FinalReceiveState

FinalSendState

GatewayConnectionError

LightningClientStateMachines

LightningInvoice

LightningOperationMeta

ReceiveError

ReceiveState

The high-level state of receiving a payment over lightning.

SelectGatewayError

SendPaymentError

SendState

The high-level state of sending a payment over lightning.

Constants

CONTRACT_CONFIRMATION_BUFFER

A two hour buffer in case either the client or gateway go offline

EXPIRATION_DELTA_LIMIT_DEFAULT

Number of blocks until outgoing lightning contracts times out and user client can refund it unilaterally

INVOICE_EXPIRATION_SECONDS_DEFAULT

Default expiration time for lightning invoices

Functions

generate_ephemeral_tweak 🔒

Type Aliases

ReceiveResult

SendResult