Crate fedimint_lnv2_client

Re-exports

pub use fedimint_lnv2_common as common;

Modules

api 🔒

cli 🔒

db 🔒

events

receive_sm 🔒

send_sm 🔒

Structs

LightningClientContext

LightningClientInit

LightningClientModule

LnurlReceiveOperationMeta

ReceiveOperationMeta

SendOperationMeta

Enums

FinalReceiveOperationState

The final state of an operation receiving a payment over lightning.

FinalSendOperationState

The final state of an operation sending a payment over lightning.

GenerateLnurlError

LightningClientStateMachines

LightningOperationMeta

ListGatewaysError

ReceiveError

ReceiveOperationState

The state of an operation receiving a payment over lightning.

RoutingInfoError

SelectGatewayError

SendOperationState

The state of an operation sending a payment over lightning.

SendPaymentError

Constants

CONTRACT_CONFIRMATION_BUFFER 🔒

A two hour buffer in case either the client or gateway go offline

EXPIRATION_DELTA_LIMIT 🔒

Number of blocks until outgoing lightning contracts times out and user client can refund it unilaterally

Type Aliases

ReceiveResult

SendResult